+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 - Show upgrade warnings in popup and log
 - Add additional check to deal with legacy gpkgs created by schematisation editor.
 - New schematisation from existing file - move migration logic to after users have clicked Create schematisation (#671)
+- Fix upgrading of schematisation before uploading
 
 
 3.17 (2025-04-02)