+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 3.18 (unreleased)
 -----------------
 
+- Show upgrade warnings in popup and log
 - Add additional check to deal with legacy gpkgs created by schematisation editor.
 - New schematisation from existing file - move migration logic to after users have clicked Create schematisation (#671)
 - Fix upgrading of schematisation before uploading